Improved 'copy if config isn't present'

+++ Dockerfiles/nginx-proxy/copy/entrypoint.sh
@@ -1,4 +1,6 @@
 #! /bin/bash
-cp -nr /defaults/* /etc/nginx
+if [ ! -e /etc/nginx/nginx.conf ] ; then
+	cp -nr /defaults/* /etc/nginx
+fi
 
 su-exec proxy:proxy nginx -g daemon off
